requirements.txt
requests
beautifulsoup4
pdfkit
lxml
langchain==0.2.0
langchain-community==0.2.0
langchain-experimental==0.0.59
langsmith
langgraph==0.0.60
llama-index
llama-parse
fastembed
chromadb
python-dotenv
langchain-groq
chainlit
fastembed
unstructured[md]
python-dotenv
python-telegram-bot
crewai
'crewai[tools]'
yfinance